# sishotel/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SisHotel",
    "description": "API REST para gestão de reservas de hotel (check-in/check-out). Arquitetura em 3 camadas: Controller → Service → Repository. FIAP - 3ESPR - CP2.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/rooms": {
      "get": {
        "tags": [
          "Quartos"
        ],
        "summary": "Lista todos os quartos",
        "operationId": "list_rooms_api_v1_rooms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoomResponse"
                  },
                  "type": "array",
                  "title": "Response List Rooms Api V1 Rooms Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Quartos"
        ],
        "summary": "Cadastra novo quarto",
        "operationId": "create_room_api_v1_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rooms/{room_id}": {
      "get": {
        "tags": [
          "Quartos"
        ],
        "summary": "Busca quarto por ID",
        "operationId": "get_room_api_v1_rooms__room_id__get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Quartos"
        ],
        "summary": "Atualiza dados do quarto",
        "operationId": "update_room_api_v1_rooms__room_id__put",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Quartos"
        ],
        "summary": "Remove um quarto (se não tiver reservas ativas)",
        "operationId": "delete_room_api_v1_rooms__room_id__delete",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Código identificador do erro"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem explicativa"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Detalhes opcionais (ex.: campos inválidos)"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoomCreate": {
        "properties": {
          "number": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Number",
            "description": "Número único do quarto"
          },
          "type": {
            "$ref": "#/components/schemas/RoomType",
            "description": "Tipo: STANDARD, DELUXE ou SUITE"
          },
          "capacity": {
            "type": "integer",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Capacity",
            "description": "Capacidade máxima de hóspedes"
          },
          "price_per_night": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Price Per Night"
          }
        },
        "type": "object",
        "required": [
          "number",
          "type",
          "capacity",
          "price_per_night"
        ],
        "title": "RoomCreate"
      },
      "RoomResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "price_per_night": {
            "type": "string",
            "title": "Price Per Night"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "type",
          "capacity",
          "price_per_night",
          "status"
        ],
        "title": "RoomResponse"
      },
      "RoomStatus": {
        "type": "string",
        "enum": [
          "ATIVO",
          "INATIVO"
        ],
        "title": "RoomStatus"
      },
      "RoomType": {
        "type": "string",
        "enum": [
          "STANDARD",
          "DELUXE",
          "SUITE"
        ],
        "title": "RoomType"
      },
      "RoomUpdate": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomType"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "price_per_night": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Night"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RoomUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}